Split animation from background
